requirements.txt
opencv-python>=4.5.0
mss>=6.1.0
numpy>=1.21.0
pywin32>=301
pyserial>=3.5
